Refactored the ButtonsPlayFriendMode component into a folder structure
